Merged PR 108587: Change system counter refclk frequency on SVP to 125 MHz
## Why is this change being made?

System counter (refclk) base frequency on the rdfremont reference design
(on which the SVP is based) runs at 125 MHz.

This value is different from silicon and is documented in the Voyager
TRM.

Programming this to 1 GHz leads to delay miscalculations and slowdowns
for AP software.

## What is changed?

Update gtimer initialization to setup SVP refclk at 125 MHz.

## How is the change being tested?

- [ ] No test, document update only
- [x] Unit tests pass
- [x] Integration tests pass
- [x] Manual tests pass (SVP windows boot)
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request corrects the system counter refclk frequency for the SVP platform from 1 GHz to 125 MHz.
- `src/libs/runtime_init/gtimer_init.c`: Defined `SVP_REFCLK_FREQUENCY` and `SVP_SCALING_FACTOR`, and added logic to set these values for the SVP platform.
- `tests/unit/libs/runtime_init/gtimer_init/test_gtimer_init.cpp`: Added new test cases to validate the SVP configuration with the updated frequency.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #2411309