Merged PR 114299: MCP - In-Band Thread Priorities
## Why is this change being made?
The MTS thread processing (on the MCP) of incoming AP requests is functional but delayed. The main processing on the MCP (at the moment) is the Power Telemetry Service, which is reading data from the SCF RAM every 10ms.

## What is changed?
- Changed Power Telemetry Service thread priority to 5
- Changed MTS thread priority to 3
- T32 script typo + timeout increase

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics) - Tested with T32 scripts from the AP
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [x] FPGA (Add FPGA Config) - Single Die DH7 R22
- [ ] Silicon
- [ ] Others, explain:

## Notes

More performance detail needs to be captured on the MCP to further understand performance.

## Risks

Script updates and IB thread pri updates

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses thread priority issues and fixes a bug related to delayed and non-responsive AP messages.
- Changes in `send_dcp_message.cmm` to update sequence number initialization for testing.
- Changes in `exec_tlm_cmpnt.c` to adjust thread priority and preempt threshold.
- Changes in `client_get_capabilities.cmm` to correct a bit position in the printed capabilities.
- Changes in `mts_init.c` to modify thread priority settings.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #2494431