Merged PR 115728: Adding checks for power_hw_gpio_connected(), and if not fake the AVS to prevent looping
## Why is this change being made?
If the system doesn't have VRs, then we need to still complete the loop.  Otherwise it will continue to try and cause issues.
## What is changed?
Checking the platform, and if no VRs then don't keep trying to read/write to the VRs.
## How is the change being tested?

- [ ] No test, document update only
- [X] Unit tests pass
- [ ] Integration tests pass
- [ ] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [ X] SVP
- [X ] FPGA (Add FPGA Config) Single Die w/o VRs.
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

Adding checks for power_hw_gpio_connected(), and if not fake the AVS.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses a bug where power control loops continue even after retries are exhausted, causing the idle queue to not process and potentially leading to system crashes.
- `src/services/power/src/power_hw_int_vrs.c`: Added checks for `power_hw_gpio_connected()` and implemented logic to fake AVS responses to prevent system crashes when GPIO is not connected.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #2523241