Merged PR 117071: Adding quiesce operation to end-to-end warm reset flow
## Why is this change being made?
This PR will separate out the shutdown from quiesce operation and have a more logical approach to the shutdown sequence

## What is changed?
Separate out the cases for sos_shutdown and sos_quiesce and have an entry point to sos_quiesce at the receive callback for PrepForCoreReset cmd from HSP

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces a quiesce operation to the end-to-end warm reset flow, allowing the system to pause operations and send an acknowledgment before a core reset. This change is crucial for handling scenarios where components like SDM/CDED need to undergo a reset.
- Added `ssi_quiesce` function in `src/services/startup_shutdown/startup_shutdown_ssi.c` to handle quiesce requests.
- Implemented `sos_notify_ssi_quiesce` and `sos_queue_quiesce` in `src/services/startup_shutdown/startup_shutdown_thread.c` to manage quiesce notifications and queuing.
- Updated `ap_core_dispatch` in `src/services/ap_core/ap_core.c` to handle `SSI_QUIESCE_ASYNC` requests.
- Modified unit tests in `tests/unit/services/startup_shutdown` to include quiesce operation testing.
- Introduced `quiesce_complete_notify` callback in `src/services/startup_shutdown/startup_shutdown_core_comm.c` for quiesce completion handling.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #2407416