Merged PR 119240: Update the CortexM7 Repo to new tag
## Why is this change being made?
Pick latest change needed for memory barrier implementation
## What is changed?
Update the CortexM7 repo tag in .xml
## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (Boot test on SVP 1D/2D and FPGA 1D)
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [x] FPGA (1D)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

Update the CortexM7 Repo to new tag

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a dependency update that revises the CortexM7 repository tag.

#### PR Summary
The update changes the CortexM7 repo version in the packages configuration to align with a newer release.
- `tools/packages.xml`: Updated the CortexM7 tag from "kingsgate_cortex_m7_v0.1.1.47" to "kingsgate_cortex_m7_v0.1.1.49_dev".
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #2603842