Merged PR 128328: Updated silibs tag to promote_silibs_sbu_a0_71
## Why is this change being made?
Pull in fix for ITB disabled config

## What is changed?
Updated Silibs tag to latest

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [ ] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

Updated silibs tag to promote_silibs_sbu_a0_71

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a configuration update to promote a newer version of the silibs repository.

#### PR Summary
The pull request updates the silibs repository tag in the packages configuration file, ensuring the project uses the updated release in its dependency management.
- `tools/packages.xml`: Changed the tag for the silibs repository from "promote_silibs_sbu_a0_69" to "promote_silibs_sbu_a0_71".
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->