Merged PR 130548: Workaround PCIe phy aux_clk deassertion failures
## Why is this change being made?

Sporadic failures are being observed where PCIe phys are continuing to
report aux_clk_active after RPSS initialization. This is indicative of
the PHY not switching to its core clock correctly, without which it
cannot be considered fully operational.

This further leads to the silibs API `pciess_rps_ready` returning a
failure when it checks for RP readiness and SCP boot crashes.

The actual root cause for aux_clk_active never going low in some
instances is still being investigated but in the meanwhile, we are
working around this issue by yielding the PCIe worker thread between the
stage where an RPSS is brought out of reset and the PHY is programmed.

This seems to avoid hitting the failures mentioned above.

## What is changed?

Yield the RPSS worker thread(s) for 500 ticks between
`send_sync_rpss_initial_config` and `send_sync_rpss_pre_rp_init_request`.

## How is the change being tested?

- [ ] No test, document update only
- [x] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (Cloudhost boot on 2d silicon g10-b2)
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [x] Silicon
- [ ] Others, explain:

## Notes

## Risks

Related work items: #2777075