Merged PR 139401: Add PCIe BDAT support
## Why is this change being made?

Populate a shared buffer with PCIe BDAT information for UEFI to consume.

## What is changed?

Add a call to publish BDAT information that contains link quality
information for all PCIe lanes on an enabled root port.

UEFI will pick this data up and add it to the BDAT ACPI table.

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [x] Silicon (2d OS boot)
- [ ] Others, explain:

## Notes

## Risks

Related work items: #2598956