Merged PR 143299: feat: add a git diff check after builds
## Why is this change being made? What is changed?
Adds a `git status` check after the build. To ensure that a PR is not pushed with latent auto-formatting pending.

## How is the change being tested?

- [ ] No test, document update only
- [X] Unit tests pass
- [X] Integration tests pass
- [ ] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [X] SVP
- [ ] FPGA (Add FPGA Config)
- [X] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature that adds a git status check to ensure the working directory remains clean after builds.

#### PR Summary
The changes add a verification step in the build and unittest pipelines to fail the build if local changes are detected.
- In `build.yaml`, a git status check is inserted after the build command to error out when uncommitted changes exist.
- In `unittest.yaml`, a similar git status check is added to enforce a clean working directory post-build.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #2938337