Merged PR 143275: cleanup: route MCP Die1 UART to UART5
## Why is this change being made? What is changed?
On the latest IFWIs, we're unable to route Die 1 accelerator UARTs to any output: we're not able to run afm commands on Die 1. Hence routing MCP Die 1 to UART5 to enable use of `gpio uart_afm` on Die 1.

## How is the change being tested?

- [ ] No test, document update only
- [X] Unit tests pass
- [X] Integration tests pass
- [X] Manual tests pass: Test on Silicon to ensure logs are updated, and MCP die 1 is available in UART5
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [X] Silicon
- [ ] Others, explain:

## Notes

## Risks

----
#### AI description  (iteration 1)
#### PR Classification
This PR is a cleanup update to align AFM mapping configurations with the new MCTP implementation.

#### PR Summary
The changes remove outdated commented code and update default UART mapping values to ensure proper MCTP routing for both dies.
- In `config/knobs_config/kng_mscp_gpio_config_knobs.xml`, cleaned up obsolete comments and updated default configurations for `uart_die_sel` and UART AFM knobs for die0 and die1.
- In `src/services/cli_gpio/gpio_cli_main.c`, revised the UART AFM mapping table to replace MCP with MCTP routing values, ensuring consistency with the updated configuration.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->