Merged PR 143056: Change log if crash dump is empty
## Why is this change being made?
KNG_E_NOT_READY is not a failure if no crash dump available, but shown as a failure

## What is changed?
Change log if crash dump is not available to transfer

## How is the change being tested?

- [ ] No test, document update only
- [x] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [ ] FPGA (Add FPGA Config)
- [x] Silicon
- [ ] Others, explain:

## Notes

## Risks

KNG_E_NOT_READY is not a failure

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix – the changes update the logging behavior in the crash dump transfer callback to correctly handle an empty crash dump scenario.

#### PR Summary
This PR enhances the crash dump transfer functionality by adding a condition to check for an empty crash dump and log a specific message when encountered.
- `src/libs/crash_dump/src/crash_dump_icc.c`: Added a conditional branch for the `KNG_E_NOT_READY` status to log "Crash dump is empty."
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #2891120