Merged PR 154510: Invalidate Cache  fix for Potential Phy corruption
## Why is this change being made?
DDR is using the PCIE PHY load area for some purpose. After that when the PCIE PHY is loaded and written to PHY we see that the PHY is corrupted.  The reason is attributed to corruption due to cached values .
A reasons to support that :
1. All RPSS are corrupted
2. looking at 2 PHYs RPSS0/1 PHY0 , the corruption seems to be the same
Both  these point to the fact that we will need to invalidate the cache.

## What is changed?
Invalidate the cache before PHY programming starts

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [x] Silicon
- [ ] Others, explain:

## Notes

## Risks

Invalidate Cache  fix for Potential Phy corruption

Related work items: #3150524