Merged PR 158265: PLDM: MCP boot flaky investigation
## Why is this change being made?

In the previous setup version, the reset was being called before opening the serial channels.

This may be the culprit behind a race condition between the reset and the serial channels. The boot banner message is being written by the MCP, however: if there's a serial buffer clearing between after this write then the program flow may not end up validating the boot message.

Thus, it's necessary to open the serial channels and then call reset of the SOC, and not the inverse.

Here's a section of the log of the failed run:
```
20251206 15:08:43.790:INFO:serial_ssh_connection.py:541: [Serial][com_id 2202 device_id None] Clearing buffer
20251206 15:08:43.792:INFO:serial_ssh_connection.py:424: Starting serial session for Blade 6, Port 2202, Device None
20251206 15:08:43.795:INFO:serial_ssh_connection.py:541: [Serial][com_id 2202 device_id None] Clearing buffer
20251206 15:08:43.792:INFO:serial_ssh_connection.py:462: Writing start serial session -i 6 -p 2202
...
A TimeoutError occurred while reading a boot message from MCP: Timeout happened while waiting for "StartupShutdownSvc::LocalCoreSyncStageRemoteEnd".
```

## What is changed?

The order of some of the steps in the setup function is changed.

Now the steps are put in the following manner:
```
Step 1: DUT setup
Step 2: Open MCP channel and BMC client channel
Step 3: Set BMC UART MUX to MCP for RVP
Step 4: Perform reset based on DUT type
Step 5: wait for MCP boot complete
Step 6: Wait for PLDM and MCTPD services to be active
Step 7: Load PLDM specification data and get MCP EID
```

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [x] Integration tests pass
- [ ] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [x] Silicon
- [ ] Others, explain:

## Notes

## Risks

Related work items: #3201687